feat(tinygo): dmsg client compiles under TinyGo (wasip1/IoT) — Phase 3
The dmsg client now builds under TinyGo for the wasip1 (IoT) target:
`make tinygo-dmsg` (tinygo build -target wasip1 -no-debug -opt=z) produces a
~2.2 MB wasm binary. Build-check main in cmd/dmsg-tinygo-probe.

Clears the remaining peripheral blockers, each with the same
!tinygo-split-plus-stub (or de-tag-the-obsolete-!tinygo) pattern established in
Phase 2:

  - pkg/netutil: net.go used net.Interface.Addrs() (absent in TinyGo's net) +
    a GOOS-split DefaultNetworkInterface with no wasip1 variant. Pure helpers
    (IsPublicIP/ExtractPort/IsVirtualInterface) stay in net.go; the
    interface-enumeration + ipinfo HTTP probe move to net_native.go (!tinygo)
    with stubs in net_tinygo.go.
  - pkg/dmsg/disc: Entry.Sign / VerifySignature were !tinygo on the stale
    "encoding/json can't run under TinyGo" assumption (false since 0.41).
    De-tagged (entry_native.go → entry_sign.go) with a stdlib encoding/json
    codec shim (interface_tinygo.go) standing in for jsoniter under TinyGo.
  - pkg/dmsg/dmsg: *net.TCPConn.SetNoDelay (absent in TinyGo) wrapped in a
    build-tagged setTCPNoDelay helper (no-op on TinyGo).
  - deployment: EnvServices moved from config_native.go (!tinygo) to the
    untagged config.go so pkg/dmsg/dmsg's InitConfig can decode the embedded
    config on TinyGo too.

Adds the `tinygo-dmsg` Makefile target and updates
docs/design/tinygo-dmsg-client.md (Phase 3 done). Normal go build ./... +
pkg/dmsg/dmsg, disc, netutil, logging, deployment test suites stay green.
