+++ pkg/netutil/net.go
@@ -1,77 +1,14 @@
 // Package netutil pkg/netutil/net.go
+//
+// The TinyGo-safe (pure) helpers. Network-interface enumeration + the
+// ipinfo.io HTTP probe live in net_native.go (//go:build !tinygo); their
+// TinyGo stubs are in net_tinygo.go.
 package netutil
 
 import (
 	"fmt"
-	"io"
 	"net"
-	"net/http"
 )
-
-// LocalNetworkInterfaceIPs gets IPs of all local interfaces.
-func LocalNetworkInterfaceIPs() ([]net.IP, error) {
-	ips, _, err := localNetworkInterfaceIPs("")
-	return ips, err
-}
-
-// NetworkInterfaceIPs gets IPs of network interface with name `name`.
-func NetworkInterfaceIPs(name string) ([]net.IP, error) {
-	_, ifcIPs, err := localNetworkInterfaceIPs(name)
-	return ifcIPs, err
-}
-
-// localNetworkInterfaceIPs gets IPs of all local interfaces. Separately returns list of IPs
-// of interface `ifcName`.
-func localNetworkInterfaceIPs(ifcName string) ([]net.IP, []net.IP, error) {
-	var ifcIPs []net.IP
-
-	ifaces, err := net.Interfaces()
-	if err != nil {
-		return nil, nil, fmt.Errorf("error getting network interfaces: %w", err)
-	}
-
-	var ips []net.IP
-	for _, iface := range ifaces {
-		if iface.Flags&net.FlagUp == 0 {
-			continue // interface down
-		}
-		if iface.Flags&net.FlagLoopback != 0 {
-			continue // loopback interface
-		}
-
-		addrs, err := iface.Addrs()
-		if err != nil {
-			return nil, nil, fmt.Errorf("error getting addresses for interface %s: %w", iface.Name, err)
-		}
-
-		for _, addr := range addrs {
-			var ip net.IP
-			switch v := addr.(type) {
-			case *net.IPNet:
-				ip = v.IP
-			case *net.IPAddr:
-				ip = v.IP
-			}
-
-			if ip == nil || ip.IsLoopback() {
-				continue
-			}
-
-			ip = ip.To4()
-			if ip == nil {
-				continue // not an ipv4 address
-			}
-
-			ips = append(ips, ip)
-
-			if ifcName != "" && iface.Name == ifcName {
-				ifcIPs = append(ifcIPs, ip)
-			}
-		}
-	}
-
-	return ips, ifcIPs, nil
-}
 
 // IsPublicIP returns true if the provided IP is public.
 // Obtained from: https://stackoverflow.com/questions/41670155/get-public-ip-in-golang
@@ -94,34 +31,6 @@ func IsPublicIP(IP net.IP) bool {
 	return false
 }
 
-// DefaultNetworkInterfaceIPs returns IP addresses for the default network interface
-func DefaultNetworkInterfaceIPs() ([]net.IP, error) {
-	networkIfc, err := DefaultNetworkInterface()
-	if err != nil {
-		return nil, fmt.Errorf("failed to get default network interface: %w", err)
-	}
-	localIPs, err := NetworkInterfaceIPs(networkIfc)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get IPs of %s: %w", networkIfc, err)
-	}
-	return localIPs, nil
-}
-
-// HasPublicIP returns true if this machine has at least one
-// publically available IP address
-func HasPublicIP() (bool, error) {
-	localIPs, err := LocalNetworkInterfaceIPs()
-	if err != nil {
-		return false, err
-	}
-	for _, IP := range localIPs {
-		if IsPublicIP(IP) {
-			return true, nil
-		}
-	}
-	return false, nil
-}
-
 // ExtractPort returns port of the given UDP or TCP address
 func ExtractPort(addr net.Addr) (uint16, error) {
 	switch address := addr.(type) {
@@ -136,46 +45,6 @@ func ExtractPort(addr net.Addr) (uint16, error) {
 	}
 }
 
-// LocalAddresses returns a list of all local addresses
-func LocalAddresses() ([]string, error) {
-	result := make([]string, 0)
-
-	ifaces, err := net.Interfaces()
-	if err != nil {
-		return nil, err
-	}
-
-	for _, iface := range ifaces {
-		// Skip loopback and down interfaces
-		if iface.Flags&net.FlagLoopback != 0 || iface.Flags&net.FlagUp == 0 {
-			continue
-		}
-		// Skip Docker/container bridge interfaces
-		if IsVirtualInterface(iface.Name) {
-			continue
-		}
-
-		addrs, err := iface.Addrs()
-		if err != nil {
-			continue
-		}
-		for _, addr := range addrs {
-			var ip net.IP
-			switch v := addr.(type) {
-			case *net.IPNet:
-				ip = v.IP
-			case *net.IPAddr:
-				ip = v.IP
-			}
-			if ip != nil && ip.IsGlobalUnicast() {
-				result = append(result, ip.String())
-			}
-		}
-	}
-
-	return result, nil
-}
-
 // IsVirtualInterface returns true for Docker bridges, veth pairs, and other
 // virtual interfaces that shouldn't be registered with the address resolver.
 func IsVirtualInterface(name string) bool {
@@ -184,22 +53,6 @@ func IsVirtualInterface(name string) bool {
 		if len(name) >= len(p) && name[:len(p)] == p {
 			return true
 		}
-	}
-	return false
-}
-
-// LocalProtocol check a condition to use dmsghttp or direct url
-func LocalProtocol() bool {
-	resp, err := http.Get("https://ipinfo.io/country")
-	if err != nil {
-		return false
-	}
-	respBody, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return false
-	}
-	if string(respBody)[:2] == "CN" {
-		return true
 	}
 	return false
 }
